R2: fix: guard initState against reload flicker; extract shared test harness; document 100ms screenshot pump
