+++ test/features/dashboard/screenshots/projects_bottom_sheet_screenshot_test.dart
@@ -1,102 +1,39 @@
-import 'package:construculator/features/dashboard/dashboard_module.dart';
-import 'package:construculator/features/dashboard/presentation/bloc/project_dropdown_bloc/project_dropdown_bloc.dart';
 import 'package:construculator/features/dashboard/presentation/widgets/projects_bottom_sheet.dart';
-import 'package:construculator/l10n/generated/app_localizations.dart';
 import 'package:construculator/libraries/project/domain/entities/enums.dart';
 import 'package:construculator/libraries/project/domain/entities/project_entity.dart';
-import 'package:construculator/libraries/project/domain/repositories/project_repository.dart';
-import 'package:construculator/libraries/project/testing/fake_project_repository.dart';
-import 'package:construculator/libraries/supabase/interfaces/supabase_wrapper.dart';
-import 'package:construculator/libraries/supabase/testing/fake_supabase_user.dart';
-import 'package:construculator/libraries/supabase/testing/fake_supabase_wrapper.dart';
-import 'package:construculator/libraries/time/testing/fake_clock_impl.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_modular/flutter_modular.dart';
 import 'package:flutter_test/flutter_test.dart';
 
-import '../../../utils/fake_app_bootstrap_factory.dart';
-import '../../../utils/screenshot/font_loader.dart';
+import '../projects_bottom_sheet_test_harness.dart';
 
 void main() {
   final size = const Size(390, 660);
   const ratio = 1.0;
   TestWidgetsFlutterBinding.ensureInitialized();
 
-  late FakeClockImpl clock;
-  late FakeSupabaseWrapper fakeSupabase;
-  late FakeProjectRepository fakeRepository;
-  late ProjectDropdownBloc bloc;
-
-  const String testUserId = 'user-1';
+  final harness = ProjectsBottomSheetTestHarness();
 
   Project buildProject({required String id, required String projectName}) {
     return Project(
       id: id,
       projectName: projectName,
-      creatorUserId: testUserId,
+      creatorUserId: ProjectsBottomSheetTestHarness.testUserId,
       createdAt: DateTime(2025, 1, 1),
       updatedAt: DateTime(2025, 4, 29, 18, 11),
       status: ProjectStatus.active,
     );
   }
 
-  Widget buildTestApp() {
-    return MaterialApp(
-      theme: createTestTheme(),
-      locale: const Locale('en'),
-      localizationsDelegates: AppLocalizations.localizationsDelegates,
-      supportedLocales: AppLocalizations.supportedLocales,
-      home: const Scaffold(body: ProjectsBottomSheet()),
-    );
-  }
-
-  Future<void> pumpSheet(WidgetTester tester) async {
-    await tester.pumpWidget(buildTestApp());
-    await tester.pump();
-    await tester.runAsync(() async {
-      await bloc.stream.firstWhere(
-        (state) =>
-            state is ProjectDropdownLoadSuccess ||
-            state is ProjectDropdownLoadFailure,
-      );
-    });
-    await tester.pump(const Duration(milliseconds: 100));
-  }
-
   setUpAll(() async {
-    await loadAppFontsAll();
-    clock = FakeClockImpl(DateTime(2025, 1, 1, 8, 0));
+    await harness.setUpAll();
   });
 
   setUp(() {
-    fakeSupabase = FakeSupabaseWrapper(clock: clock);
-    fakeRepository = FakeProjectRepository();
-
-    final bootstrap = FakeAppBootstrapFactory.create(
-      supabaseWrapper: fakeSupabase,
-    );
-    Modular.init(DashboardModule(bootstrap));
-    Modular.replaceInstance<SupabaseWrapper>(fakeSupabase);
-    Modular.replaceInstance<ProjectRepository>(fakeRepository);
-
-    fakeSupabase.setCurrentUser(
-      FakeUser(
-        id: testUserId,
-        email: 'user-1@example.com',
-        createdAt: clock.now().toIso8601String(),
-        appMetadata: const {},
-        userMetadata: const {},
-      ),
-    );
-
-    bloc = Modular.get<ProjectDropdownBloc>();
-    Modular.replaceInstance<ProjectDropdownBloc>(bloc);
+    harness.setUp();
   });
 
   tearDown(() {
-    bloc.close();
-    fakeRepository.dispose();
-    Modular.destroy();
+    harness.tearDown();
   });
 
   group('ProjectsBottomSheet Screenshot Tests', () {
@@ -105,13 +42,17 @@ void main() {
       tester.view.devicePixelRatio = ratio;
       addTearDown(tester.view.reset);
 
-      fakeRepository.setAccessibleProjects([
+      harness.fakeRepository.setAccessibleProjects([
         buildProject(id: 'project-1', projectName: 'My project'),
         buildProject(id: 'project-2', projectName: 'Material of building'),
         buildProject(id: 'project-3', projectName: 'MD bungalow'),
       ]);
 
-      await pumpSheet(tester);
+      // Extra 100ms pump lets the golden render stabilise before capture.
+      await harness.pumpSheet(
+        tester,
+        extraPump: const Duration(milliseconds: 100),
+      );
 
       await expectLater(
         find.byType(ProjectsBottomSheet),
@@ -126,9 +67,12 @@ void main() {
       tester.view.devicePixelRatio = ratio;
       addTearDown(tester.view.reset);
 
-      fakeRepository.setAccessibleProjects([]);
+      harness.fakeRepository.setAccessibleProjects([]);
 
-      await pumpSheet(tester);
+      await harness.pumpSheet(
+        tester,
+        extraPump: const Duration(milliseconds: 100),
+      );
 
       await expectLater(
         find.byType(ProjectsBottomSheet),
